Implemment _get_spatial_information for the tiled_full case
